# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "XYZW Web Helper Hardening Contracts",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/wechat-proxy/qrstatus": {
      "post": {
        "operationId": "wechatQrStatus",
        "summary": "轮询微信二维码状态",
        "tags": [
          "Proxy"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "uuid": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  }
                },
                "required": [
                  "uuid"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "string"
                }
              },
              "text/html": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wechat-proxy/hortor-login": {
      "post": {
        "operationId": "wechatHortorLogin",
        "summary": "代理转发 Hortor 登录请求",
        "tags": [
          "Proxy"
        ],
        "parameters": [
          {
            "name": "x-xyzw-device-unique-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[\\w.:-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "string"
                }
              },
              "application/*+json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "string"
                }
              },
              "text/plain": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}